Basic before / after.
It's late and I'm tired, but this needs to get committed and it can get
fixed later. There are is missing functionality and test cases for
failures in after and in the test case, but these will have to come
later. Also, these tests are really messy, and refactoring /
de-duplication is needed all-around.